Clean out the old solver source files
